// src/types/responses.ts
export interface ApiResponse<T> {
  success: boolean;
  message: string;
  data: T;
  code: string;
}



export interface AuthData {
  access_token: string;
  expires_in: number;
  refresh_expires_in: number;
  refresh_token: string;
  token_type: string;
  "not-before-policy": number;
  session_state: string;
  scope: string;
}